Move recipe actions to namespace to make space for inventory actions
