+++ resources/views/pages/recipes/⚡edit.blade.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Actions\ImportRecipeFromUrl;
+use App\Actions\Recipes\ImportRecipeFromUrl;
 use App\Livewire\Forms\Recipes\RecipeForm;
 use App\Models\Recipe;
 use Livewire\Component;
